fix : 공통 예외 응답이 우선되는 문제 해결
