--- src/main/java/makeus/cmc/malmo/adaptor/in/web/exception/CommonExceptionHandler.java
@@ -1,46 +1,0 @@
-package makeus.cmc.malmo.adaptor.in.web.exception;
-
-import lombok.extern.slf4j.Slf4j;
-import makeus.cmc.malmo.adaptor.out.oidc.exception.OidcIdTokenException;
-import makeus.cmc.malmo.adaptor.out.persistence.exception.MemberNotFoundException;
-import makeus.cmc.malmo.application.exception.InvalidRefreshTokenException;
-import org.hibernate.TypeMismatchException;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.HttpRequestMethodNotSupportedException;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.NoHandlerFoundException;
-
-import java.io.IOException;
-import java.util.NoSuchElementException;
-
-@Slf4j
-@RestControllerAdvice
-public class CommonExceptionHandler {
-
-    @ExceptionHandler({NoHandlerFoundException.class, TypeMismatchException.class})
-    public ResponseEntity<ErrorResponse> handleBadRequestException(Exception e) {
-        log.error("[CommonExceptionHandler: handleBadRequestException 호출]", e);
-        return ErrorResponse.of(ErrorCode.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(HttpRequestMethodNotSupportedException.class)
-    public ResponseEntity<ErrorResponse> handleHttpRequestMethodNotSupportedException(HttpRequestMethodNotSupportedException e) {
-        log.error("[CommonExceptionHandler: handleHttpRequestMethodNotSupportedException 호출]", e);
-        return ErrorResponse.of(ErrorCode.METHOD_NOT_ALLOWED);
-    }
-
-    @ExceptionHandler({IllegalArgumentException.class, IOException.class})
-    public ResponseEntity<ErrorResponse> handleInternalServerException(Exception e) {
-        log.error("[CommonExceptionHandler: handleInternalServerException 호출]", e);
-        return ErrorResponse.of(ErrorCode.INTERNAL_SERVER_ERROR);
-    }
-
-    @ExceptionHandler(RuntimeException.class)
-    public ResponseEntity<ErrorResponse> handleRuntimeException(Exception e) {
-        log.error("[CommonExceptionHandler: handleRuntimeException 호출]", e);
-        return ErrorResponse.of(ErrorCode.INTERNAL_SERVER_ERROR);
-    }
-
-}
